Started on Wator objects

+++ src/cellsociety_team12/simulations/Simulation.java
@@ -39,8 +39,8 @@ import javafx.util.Duration;
  */
 public abstract class Simulation extends Application {
 
-    private static final int FRAMES_PER_SECOND = 1;
-    private static final int MILLISECOND_DELAY = 1000 / FRAMES_PER_SECOND;
+    private static final double FRAMES_PER_SECOND = 1;
+    private static final double MILLISECOND_DELAY = 1000 / FRAMES_PER_SECOND;
     private static final int GRID_SIZEX = 400;
     private static final int GRID_SIZEY = 400;
     private static final int GRID_XLOC = 100;
